+++ src/OpenSearch/Common/Exceptions/Serializer/JsonErrorException.php
@@ -21,69 +21,14 @@ declare(strict_types=1);
 
 namespace OpenSearch\Common\Exceptions\Serializer;
 
-use OpenSearch\Exception\OpenSearchExceptionInterface;
+// @phpstan-ignore classConstant.deprecatedClass
+@trigger_error(JsonErrorException::class . ' is deprecated in 2.4.0 and will be removed in 3.0.0. Use \OpenSearch\Exception\JsonErrorException instead.', E_USER_DEPRECATED);
 
 /**
  * Class JsonErrorException
+ *
+ * @deprecated in 2.4.0 and will be removed in 3.0.0. Use \OpenSearch\Exception\JsonErrorException instead.
  */
-class JsonErrorException extends \Exception implements OpenSearchExceptionInterface
+class JsonErrorException extends \OpenSearch\Exception\JsonException
 {
-    /**
-     * @var mixed
-     */
-    private $input;
-
-    /**
-     * @var mixed
-     */
-    private $result;
-
-    /**
-     * @var string[]
-     */
-    private static $messages = array(
-        JSON_ERROR_DEPTH => 'The maximum stack depth has been exceeded',
-        JSON_ERROR_STATE_MISMATCH => 'Invalid or malformed JSON',
-        JSON_ERROR_CTRL_CHAR => 'Control character error, possibly incorrectly encoded',
-        JSON_ERROR_SYNTAX => 'Syntax error',
-        JSON_ERROR_UTF8 => 'Malformed UTF-8 characters, possibly incorrectly encoded',
-        JSON_ERROR_RECURSION => 'One or more recursive references in the value to be encoded',
-        JSON_ERROR_INF_OR_NAN => 'One or more NAN or INF values in the value to be encoded',
-        JSON_ERROR_UNSUPPORTED_TYPE => 'A value of a type that cannot be encoded was given',
-
-        // JSON_ERROR_* constant values that are available on PHP >= 7.0
-        9 => 'Decoding of value would result in invalid PHP property name', //JSON_ERROR_INVALID_PROPERTY_NAME
-        10 => 'Attempted to decode nonexistent UTF-16 code-point' //JSON_ERROR_UTF16
-    );
-
-    /**
-     * @param mixed $input
-     * @param mixed $result
-     */
-    public function __construct(int $code, $input, $result, ?\Throwable $previous = null)
-    {
-        if (isset(self::$messages[$code]) !== true) {
-            throw new \InvalidArgumentException(sprintf('Encountered unknown JSON error code: [%d]', $code));
-        }
-
-        parent::__construct(self::$messages[$code], $code, $previous);
-        $this->input = $input;
-        $this->result = $result;
-    }
-
-    /**
-     * @return mixed
-     */
-    public function getInput()
-    {
-        return $this->input;
-    }
-
-    /**
-     * @return mixed
-     */
-    public function getResult()
-    {
-        return $this->result;
-    }
 }