+++ src/OpenSearch/Common/Exceptions/UnexpectedValueException.php
@@ -21,8 +21,12 @@ declare(strict_types=1);
 
 namespace OpenSearch\Common\Exceptions;
 
-use OpenSearch\Exception\OpenSearchExceptionInterface;
+// @phpstan-ignore classConstant.deprecatedClass
+@trigger_error(UnexpectedValueException::class . ' is deprecated in 2.4.0 and will be removed in 3.0.0. Use \OpenSearch\Exception\UnexpectedValueException instead.', E_USER_DEPRECATED);
 
-class UnexpectedValueException extends \UnexpectedValueException implements OpenSearchExceptionInterface
+/**
+ * @deprecated in 2.4.0 and will be removed in 3.0.0. Use \OpenSearch\Exception\UnexpectedValueException instead.
+ */
+class UnexpectedValueException extends \OpenSearch\Exception\UnexpectedValueException
 {
 }