+++ src/OpenSearch/Endpoints/AsyncSearch/Delete.php
@@ -21,7 +21,7 @@ declare(strict_types=1);
 
 namespace OpenSearch\Endpoints\AsyncSearch;
 
-use OpenSearch\Common\Exceptions\RuntimeException;
+use OpenSearch\Exception\RuntimeException;
 use OpenSearch\Endpoints\AbstractEndpoint;
 
 class Delete extends AbstractEndpoint