+++ src/OpenSearch/Endpoints/SearchableSnapshots/RepositoryStats.php
@@ -21,7 +21,7 @@ declare(strict_types=1);
 
 namespace OpenSearch\Endpoints\SearchableSnapshots;
 
-use OpenSearch\Common\Exceptions\RuntimeException;
+use OpenSearch\Exception\RuntimeException;
 use OpenSearch\Endpoints\AbstractEndpoint;
 
 class RepositoryStats extends AbstractEndpoint